+++ src/Transports/RabbitMQ/Wolverine.RabbitMQ/Internal/RabbitMqQueue.Bindings.cs
@@ -1,85 +1,5 @@
 ﻿namespace Wolverine.RabbitMQ.Internal;
 
-public class TopicBindingExchange
-{
-    private readonly RabbitMqTransport _transport;
-    private readonly string _topicPattern;
-    private readonly string _exchangeName;
-
-    public TopicBindingExchange(RabbitMqTransport transport, string topicPattern, string exchangeName)
-    {
-        _transport = transport;
-        _topicPattern = topicPattern;
-        _exchangeName = exchangeName;
-    }
-
-    /// <summary>
-    ///     Create a binding of the topic pattern previously specified to a Rabbit Mq queue
-    /// </summary>
-    /// <param name="queueName">The name of the Rabbit Mq queue</param>
-    /// <param name="configureQueue">Optionally configure </param>
-    public void ToQueue(string queueName, Action<IRabbitMqQueue>? configureQueue = null)
-    {
-        var queue = _transport.Queues[queueName];
-        var binding = queue.BindExchange(_exchangeName, _topicPattern);
-        configureQueue?.Invoke(binding.Queue);
-    }
-}
-
-
-// Legacy interface for exchange -> queue binding expressions
-public interface IRabbitMqBindableExchange : IRabbitMqExchange
-{
-    public TopicBindingExchange BindTopic(string topicPattern);
-
-    public RabbitMqBinding BindQueue(string queueName, string? bindingKey = null);
-    
-}
-
-public class RabbitMqExchangeConfigurationExpression : IRabbitMqBindableExchange
-{
-    private readonly RabbitMqExchange _exchange;
-    private readonly RabbitMqTransport _transport;
-
-    public RabbitMqExchangeConfigurationExpression(string exchangeName, RabbitMqTransport transport)
-    {
-        _exchange = transport.Exchanges[exchangeName];
-        _transport = transport;
-    }
-
-    public string Name => _exchange.Name;
-
-    public bool IsDurable
-    {
-        get => _exchange.IsDurable;
-        set => _exchange.IsDurable = value;
-    }
-    public ExchangeType ExchangeType
-    {
-        get => _exchange.ExchangeType;
-        set => _exchange.ExchangeType = value;
-    }
-    public bool AutoDelete
-    {
-        get => _exchange.AutoDelete;
-        set => _exchange.AutoDelete = value;
-    }
-    public IDictionary<string, object> Arguments => _exchange.Arguments;
-    public TopicBindingExchange BindTopic(string topicPattern)
-    {
-        var queue = _transport.Queues[Name];
-        queue.BindTopic(topicPattern);
-        return new TopicBindingExchange(_transport, topicPattern, Name);
-    }
-
-    public RabbitMqBinding BindQueue(string queueName, string? bindingKey = null)
-    {
-        var queue = _transport.Queues[queueName];
-        return queue.BindExchange(Name, bindingKey);
-    }
-}
-
-
 public partial class RabbitMqQueue
 {
     private readonly List<RabbitMqBinding> _bindings = [];
